# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/invoices": {
      "get": {
        "description": "Retrieves a list of all invoices in the system",
        "operationId": "InvoicesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of invoices retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvoiceDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all invoices",
        "tags": [
          "invoices"
        ]
      },
      "post": {
        "description": "Creates a new invoice in the system",
        "operationId": "InvoicesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Invoice created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        },
        "summary": "Create invoice",
        "tags": [
          "invoices"
        ]
      }
    },
    "/invoices/alt/{altId}": {
      "get": {
        "description": "Retrieves an invoice by its alternative ID",
        "operationId": "InvoicesController_findByAltId",
        "parameters": [
          {
            "name": "altId",
            "required": true,
            "in": "path",
            "description": "Alternative ID of the invoice",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDto"
                }
              }
            }
          },
          "404": {
            "description": "Invoice not found"
          }
        },
        "summary": "Get invoice by alternative ID",
        "tags": [
          "invoices"
        ]
      }
    },
    "/invoices/user/{userId}": {
      "get": {
        "description": "Retrieves all invoices associated with a specific user",
        "operationId": "InvoicesController_findByUserId",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "ID of the user",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of invoices retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvoiceDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get invoices by user ID",
        "tags": [
          "invoices"
        ]
      }
    },
    "/invoices/{id}": {
      "get": {
        "description": "Retrieves an invoice by its ID",
        "operationId": "InvoicesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID of the invoice",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDto"
                }
              }
            }
          },
          "404": {
            "description": "Invoice not found"
          }
        },
        "summary": "Get invoice by ID",
        "tags": [
          "invoices"
        ]
      },
      "put": {
        "description": "Updates an existing invoice in the system",
        "operationId": "InvoicesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID of the invoice to update",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invoice updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Invoice not found"
          }
        },
        "summary": "Update invoice",
        "tags": [
          "invoices"
        ]
      },
      "delete": {
        "description": "Deletes an invoice from the system",
        "operationId": "InvoicesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID of the invoice to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice deleted successfully"
          },
          "404": {
            "description": "Invoice not found"
          }
        },
        "summary": "Delete invoice",
        "tags": [
          "invoices"
        ]
      }
    },
    "/invoices-items": {
      "get": {
        "description": "Retrieves a list of all invoice items in the system",
        "operationId": "InvoicesItemsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of invoice items retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvoiceItemDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all invoice items",
        "tags": [
          "invoices-items"
        ]
      },
      "post": {
        "description": "Creates a new association between an invoice and an item",
        "operationId": "InvoicesItemsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Invoice item created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceItemDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        },
        "summary": "Create invoice item",
        "tags": [
          "invoices-items"
        ]
      }
    },
    "/invoices-items/invoice/{invoiceId}": {
      "get": {
        "description": "Retrieves all items associated with a specific invoice",
        "operationId": "InvoicesItemsController_findByInvoiceId",
        "parameters": [
          {
            "name": "invoiceId",
            "required": true,
            "in": "path",
            "description": "ID of the invoice",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of invoice items retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvoiceItemDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get invoice items by invoice ID",
        "tags": [
          "invoices-items"
        ]
      },
      "delete": {
        "description": "Deletes all items associated with a specific invoice",
        "operationId": "InvoicesItemsController_removeByInvoiceId",
        "parameters": [
          {
            "name": "invoiceId",
            "required": true,
            "in": "path",
            "description": "ID of the invoice",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice items deleted successfully"
          }
        },
        "summary": "Delete invoice items by invoice ID",
        "tags": [
          "invoices-items"
        ]
      }
    },
    "/invoices-items/item/{itemId}": {
      "get": {
        "description": "Retrieves all invoices associated with a specific item",
        "operationId": "InvoicesItemsController_findByItemId",
        "parameters": [
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "description": "ID of the item",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of invoice items retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvoiceItemDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get invoice items by item ID",
        "tags": [
          "invoices-items"
        ]
      },
      "delete": {
        "description": "Deletes all invoices associated with a specific item",
        "operationId": "InvoicesItemsController_removeByItemId",
        "parameters": [
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "description": "ID of the item",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice items deleted successfully"
          }
        },
        "summary": "Delete invoice items by item ID",
        "tags": [
          "invoices-items"
        ]
      }
    },
    "/invoices-items/{invoiceId}/{itemId}": {
      "get": {
        "description": "Retrieves a specific invoice item by invoice ID and item ID",
        "operationId": "InvoicesItemsController_findOne",
        "parameters": [
          {
            "name": "invoiceId",
            "required": true,
            "in": "path",
            "description": "ID of the invoice",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "description": "ID of the item",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice item retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceItemDto"
                }
              }
            }
          },
          "404": {
            "description": "Invoice item not found"
          }
        },
        "summary": "Get specific invoice item",
        "tags": [
          "invoices-items"
        ]
      },
      "delete": {
        "description": "Deletes a specific association between an invoice and an item",
        "operationId": "InvoicesItemsController_remove",
        "parameters": [
          {
            "name": "invoiceId",
            "required": true,
            "in": "path",
            "description": "ID of the invoice",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "description": "ID of the item",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice item deleted successfully"
          },
          "404": {
            "description": "Invoice item not found"
          }
        },
        "summary": "Delete specific invoice item",
        "tags": [
          "invoices-items"
        ]
      }
    }
  },
  "info": {
    "title": "Inventory Service API",
    "description": "API documentation for the Inventory Service",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "InvoiceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "The unique identifier for the invoice"
          },
          "alt_id": {
            "type": "string",
            "description": "An alternative identifier for the invoice"
          },
          "total": {
            "type": "number",
            "description": "The total amount of the invoice",
            "example": 100.5
          },
          "paid": {
            "type": "boolean",
            "description": "Whether the invoice has been paid",
            "default": false
          },
          "user_id": {
            "type": "string",
            "description": "The ID of the user associated with the invoice"
          },
          "created_by": {
            "type": "string",
            "description": "Who created the invoice"
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "When the invoice was created"
          },
          "last_update": {
            "format": "date-time",
            "type": "string",
            "description": "When the invoice was last updated"
          },
          "last_changed_by": {
            "type": "string",
            "description": "Who last changed the invoice"
          }
        },
        "required": [
          "total",
          "user_id"
        ]
      },
      "CreateInvoiceDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "description": "The total amount of the invoice",
            "example": 100.5
          },
          "paid": {
            "type": "boolean",
            "description": "Whether the invoice has been paid",
            "default": false
          },
          "user_id": {
            "type": "string",
            "description": "The ID of the user associated with the invoice"
          },
          "created_by": {
            "type": "string",
            "description": "Who created the invoice"
          }
        },
        "required": [
          "total",
          "user_id"
        ]
      },
      "UpdateInvoiceDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "description": "The total amount of the invoice",
            "example": 100.5
          },
          "paid": {
            "type": "boolean",
            "description": "Whether the invoice has been paid"
          },
          "user_id": {
            "type": "string",
            "description": "The ID of the user associated with the invoice"
          },
          "last_changed_by": {
            "type": "string",
            "description": "Who last changed the invoice"
          }
        }
      },
      "InvoiceItemDto": {
        "type": "object",
        "properties": {
          "invoice_id": {
            "type": "string",
            "description": "The ID of the invoice"
          },
          "item_id": {
            "type": "string",
            "description": "The ID of the item"
          }
        },
        "required": [
          "invoice_id",
          "item_id"
        ]
      }
    }
  }
}